+++ src/matrix.h
@@ -3,10 +3,10 @@
 #include <types.h>
 
 extern int matrix_init(int outmodno);
-extern byte matrix_getx(void);
-extern byte matrix_gety(void);
-extern int matrix_set(byte x, byte y, RGB *color);
-extern int matrix_fill(byte start_x, byte start_y, byte end_x, byte end_y, RGB *color);
+extern int matrix_getx(void);
+extern int matrix_gety(void);
+extern int matrix_set(int x, int y, RGB *color);
+extern int matrix_fill(int start_x, int start_y, int end_x, int end_y, RGB *color);
 extern int matrix_clear(void);
 extern int matrix_render(void);
 extern int matrix_deinit(void);
